+++ .rubocop.yml
@@ -2,3 +2,5 @@ Documentation:
   Enabled: false
   LineLength:
     Max: 99
+  Style/FrozenStringLiteralComment:
+    Enabled: false
