+++ README.md
@@ -1,6 +1,6 @@
-# PROYECTO TRAINING CON RUBY ON RAILS [![Build Status](https://travis-ci.com/wx-inc-trainings/me-rails.svg?branch=kickoff-api)](https://travis-ci.com/wx-inc-trainings/me-rails)
+# PROYECTO TRAINING CON RUBY ON RAILS [![Build Status](https://travis-ci.com/wx-inc-trainings/me-rails.svg?branch=master)](https://travis-ci.com/wx-inc-trainings/me-rails)
 
-* (2) Integración API Externa [![Build Status](https://travis-ci.com/wx-inc-trainings/me-rails.svg?branch=service-open-library)](https://travis-ci.com/wx-inc-trainings/me-rails)
+* (2) Integración API Externa 
 
 * Database creation
