+++ app/controllers/api/open_library_controller.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+class Api::OpenLibraryController < ApplicationController
+  def show
+    @book = OpenLibraryService.get_library(params[:ISBN])
+    render json: @book[:result], status: @book[:status?]
+  end
+end
