--- app/controllers/open_library_controller.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-class OpenLibraryController < ApplicationController
-  def index
-    @book = AppServices::OpenLibraryService.get_library(params['ISBN'])
-    render json: @book[:result], status: @book[:status?]
-  end
-end
