+++ config/routes.rb
@@ -1,4 +1,7 @@
 Rails.application.routes.draw do
-  get 'api/open-library', to: 'open_library#index', at: "open_library_api_get"
+  
+  namespace :api do
+    resources :open_library, only: :show, param: :ISBN
+  end
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
 end
